Add files via upload
The run_all_fits() function in the run_from_csv.py script is designed to automate the fitting of light curves for multiple supernovae using the new_LinearR() function. It reads in a CSV file containing a list of supernova names along with start and end Modified Julian Date (MJD) values for seven photometric filters: UVW2, UVM2, UVW1, U, B, and V. For each row in the CSV, it constructs the corresponding photometry filename (e.g., SN2005CS.txt) and passes this filename, along with the associated MJD date ranges, to the new_LinearR() function. The script prints which supernova it is processing and reports whether the fit succeeded or failed.

To work correctly, the CSV file must begin with a header row containing lowercase column names in the following order: sn, uvw2start, uvw2end, uvm2start, uvm2end, uvw1start, uvw1end, ustart, uend, bstart, bend, vstart, and vend. Each subsequent row should contain the name of a supernova and its associated MJD date ranges. If any filter is not used for a particular SN, its start and end values can be set to 0. The .txt photometry files for each SN must exist in the same directory as the script and should be named exactly to match the SN name followed by .txt (e.g., SN2005CS.txt). Of course, if one wishes to use different filters then that is fine, but all of the code must be adjusted accordingly.